Add files via upload
In this notebook file I tried predicting the IC50 binding affinity of the molecules in the "split_1.csv" dataset. This proved to be inaccurate. It could be possibly improved by testing with different features however it seems more reasonable to use a different model that encompasses more of the data.