Add files via upload
This notebook file reads data from "split_1.csv" and tries to predict the IC50 binding affinity. This has shown to be inaccurate and ineffective. Perhaps with further testing of varied feature combinations it can be improved. However it is more reasonable to first consider other machine learning models to make sense of the data.